third_party/WebKit/LayoutTests/fast/forms/form-collection-elements-order-expected.txt

CONSOLE WARNING: line 27: The <keygen> element is deprecated and will be removed in M57, around March 2017. See https://www.chromestatus.com/features/5716060992962560 for more details.
This test examines the order of form associated elements which are classified as 'listed'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


- Checks the order of form associated elements.
PASS checkOrder([button, fieldset, input, keygen, object, output, select, textarea]) is "OK"

- Checks the order after some elements removed/inserted.
PASS checkOrder([button, fieldset, input, keygen, output, select, textarea]) is "OK"
PASS checkOrder([fieldset, input, keygen, output, select, textarea]) is "OK"
PASS checkOrder([object, fieldset, input, keygen, output, select, textarea]) is "OK"
PASS checkOrder([object, fieldset, input, keygen, output, select, textarea, button]) is "OK"

- Checks the order of form associated elements with form attribute.
PASS checkOrder([button, fieldset, input, keygen, object, output, select, textarea]) is "OK"

- Checks the order after some elements removed/inserted with form attribute.
PASS checkOrder([button, fieldset, input, keygen, output, select, textarea]) is "OK"
PASS checkOrder([fieldset, input, keygen, output, select, textarea]) is "OK"
PASS checkOrder([fieldset, input, keygen, output, select]) is "OK"
PASS checkOrder([fieldset, input, keygen, object, output, select]) is "OK"
PASS checkOrder([textarea, fieldset, input, keygen, object, output, select]) is "OK"
PASS checkOrder([textarea, fieldset, input, keygen, object, output, select, button]) is "OK"
PASS successfullyParsed is true

TEST COMPLETE
Test
